Rewind:
A "Rewind 10s" button is available in the control bar.
Mute/Unmute:

A Mute/Unmute toggle button (VolumeX/Volume2 icons) is in the control bar.
Playback Speed Control:

Buttons for 0.5x, 1x, and 1.5x playback speeds are present in the control bar.
Autoplay Next Video with Countdown:

When a video ends, a 10-second countdown starts automatically to play the next video in the playlist.
The countdown is displayed below the currently playing video in the playlist sidebar.
Mark Video as Completed:

Videos are marked as completed (with a CheckCircle icon) in the playlist sidebar after they finish playing

none of these are displaying in the video player